refactor(commands): move switch/remove orchestration out of main.rs
Move handle_switch_command into commands/worktree/switch.rs and
handle_remove_command (with RemovePlans / validate_remove_targets) into
a new commands/remove.rs, leaving thin dispatch arms in main.rs. Shared
helpers flag_pair and print_windows_picker_unavailable move to
commands/mod.rs. Pure code motion; behavior unchanged.

Also: allowlist remove.rs in the output-system guard (its --format=json
stdout was previously exempt only because main.rs is not scanned),
update two doc comments that pointed at main.rs's approve_remove, drop
pub from the now file-local run_switch / SwitchOptions, and import
flag_pair in merge.rs instead of six inline super:: calls.
