+++ src/commands/worktree/switch.rs
@@ -32,11 +32,12 @@ use worktrunk::styling::{
 
 use super::resolve::{compute_worktree_path, offer_bare_repo_worktree_path_fix, path_mismatch};
 use super::types::{CreationMethod, SwitchBranchInfo, SwitchPlan, SwitchResult};
-use crate::cli::SwitchFormat;
+use crate::cli::{SwitchArgs, SwitchFormat};
 use crate::commands::backup::back_up_clobbered_path_now;
 use crate::commands::command_approval::approve_hooks;
 use crate::commands::command_executor::FailureStrategy;
 use crate::commands::command_executor::{CommandContext, build_hook_context};
+use crate::commands::flag_pair;
 use crate::commands::hook_plan::{ApprovedHookPlan, HookPlanBuilder, register_planned};
 use crate::commands::hooks::{HookAnnouncer, execute_hook};
 use crate::commands::template_vars::TemplateVars;
@@ -1298,18 +1299,18 @@ fn emit_switch_json(
 }
 
 /// Options for the switch command
-pub struct SwitchOptions<'a> {
-    pub branch: &'a str,
-    pub create: bool,
-    pub base: Option<&'a str>,
-    pub execute: Option<&'a str>,
-    pub execute_args: &'a [String],
-    pub yes: bool,
-    pub clobber: bool,
+struct SwitchOptions<'a> {
+    branch: &'a str,
+    create: bool,
+    base: Option<&'a str>,
+    execute: Option<&'a str>,
+    execute_args: &'a [String],
+    yes: bool,
+    clobber: bool,
     /// Resolved from --cd/--no-cd flags: Some(true) = cd, Some(false) = no cd, None = use config
-    pub change_dir: Option<bool>,
-    pub verify: bool,
-    pub format: crate::cli::SwitchFormat,
+    change_dir: Option<bool>,
+    verify: bool,
+    format: crate::cli::SwitchFormat,
 }
 
 /// Run pre-switch hooks before branch resolution or worktree creation.
@@ -1768,7 +1769,7 @@ impl SwitchPipeline<'_> {
 }
 
 /// Handle the switch command.
-pub fn run_switch(
+fn run_switch(
     opts: SwitchOptions<'_>,
     config: &mut UserConfig,
     binary_name: &str,
@@ -1825,6 +1826,64 @@ pub fn run_switch(
         shell_integration_binary: Some(binary_name),
     }
     .run()
+}
+
+/// Entry point for the `wt switch` command.
+pub fn handle_switch_command(args: SwitchArgs, yes: bool) -> anyhow::Result<()> {
+    let verify = args.hooks.resolve();
+
+    // With no branch argument, `wt switch` opens a TUI picker — config
+    // deprecation warnings would render above the picker and push it down.
+    // They're still shown by other commands (`wt list`, `wt merge`, …).
+    if args.branch.is_none() {
+        worktrunk::config::suppress_warnings();
+    }
+
+    UserConfig::load()
+        .context("Failed to load config")
+        .and_then(|mut config| {
+            // No branch argument: open interactive picker
+            let change_dir_flag = flag_pair(args.cd, args.no_cd);
+
+            let Some(branch) = args.branch else {
+                #[cfg(unix)]
+                {
+                    return crate::commands::handle_picker(
+                        args.branches,
+                        args.remotes,
+                        change_dir_flag,
+                        args.format,
+                    );
+                }
+
+                #[cfg(not(unix))]
+                {
+                    use worktrunk::git::WorktrunkError;
+                    // Suppress unused variable warnings on Windows
+                    let _ = (args.branches, args.remotes, change_dir_flag);
+
+                    crate::commands::print_windows_picker_unavailable();
+                    return Err(WorktrunkError::AlreadyDisplayed { exit_code: 2 }.into());
+                }
+            };
+
+            run_switch(
+                SwitchOptions {
+                    branch: &branch,
+                    create: args.create,
+                    base: args.base.as_deref(),
+                    execute: args.execute.as_deref(),
+                    execute_args: &args.execute_args,
+                    yes,
+                    clobber: args.clobber,
+                    change_dir: change_dir_flag,
+                    verify,
+                    format: args.format,
+                },
+                &mut config,
+                &crate::binary_name(),
+            )
+        })
 }
 
 /// Whether `value` is a single clean program-name token — the form `--execute`
