+++ tests/integration_tests/output_system_guard.rs
@@ -63,6 +63,8 @@ const STDOUT_ALLOWED_PATHS: &[&str] = &[
     "for_each.rs",
     // JSON output for wt merge --format=json
     "merge.rs",
+    // JSON output for wt remove --format=json
+    "remove.rs",
     // Hook listing output for wt hook show (paged)
     "hook_commands.rs",
 ];
